Example high-dimensional matched case-control data

A simulated 1:5 matched case-control dataset with 20 covariates, where 10 covariates are truly non-zero. The data are split into training and test sets and include both the true underlying coefficients and an externally supplied coefficient vector for KL-based integration.

Usage

data(ExampleData_cc_highdim)

Format

A list containing:

train

List with elements y, z, and stratum.

test

Same structure as train.

beta_true

Numeric vector (length 50) of true coefficients.

beta_external

Numeric vector (length 50) representing external coefficients.

Examples

data(ExampleData_cc_highdim)
table(ExampleData_cc_highdim$train$y)
#> 
#>   0   1 
#> 450  50 
head(ExampleData_cc_highdim$train$z)
#>              Z1         Z2          Z3         Z4         Z5           Z6
#> [1,]  1.4611595 -0.4337634 -0.19580912 -0.9710716 -2.0865843 -0.239835608
#> [2,] -0.4022167 -1.0825005 -0.21833539 -0.9249490 -0.8127299  1.085952441
#> [3,] -1.0639890  0.1340951  0.28180593 -0.3163208  1.7753520 -0.080932050
#> [4,]  2.5053092 -0.7174750  0.07707777 -0.1388942 -0.2036942 -0.816647571
#> [5,] -0.4559238 -0.2339386  0.75748825 -0.5373669  0.9870436 -0.397471239
#> [6,] -0.1274066 -0.9105335 -1.60320176 -0.4957522  0.7000221  0.003717702
#>                Z7         Z8         Z9        Z10        Z11        Z12
#> [1,] -0.968512774 -0.5317932  0.5332914 -0.4253693  0.7721337 -0.2454276
#> [2,]  0.757822313 -1.6245614  0.5438503  0.4338086 -0.7681879 -0.4537053
#> [3,] -2.086653049 -0.6027954 -0.2068006 -0.2999812  0.4455223 -0.2156660
#> [4,]  0.008336971 -1.4387924 -0.1056239  0.3303660  0.3547015 -1.0132220
#> [5,]  1.328067232 -0.7843972 -1.3533522  1.0805977  1.2105810  1.7303790
#> [6,] -2.211353611 -0.2443305 -1.2150816 -1.1020840 -1.0640061 -1.3632340
#>              Z13        Z14        Z15         Z16         Z17        Z18
#> [1,]  0.05763206  0.7916978 -0.3372248 -0.24097349 -1.62260415  1.0745083
#> [2,]  0.65016454 -0.4662842  0.5785056  0.62206135 -0.01953363  0.8032010
#> [3,]  0.49218663  0.6169846 -0.2566177  1.22935275 -0.01466452 -0.4857751
#> [4,] -2.12257983 -0.7336376 -0.5649878 -0.66695983 -0.74322925 -0.7392598
#> [5,]  0.69525556 -0.9749203  1.0245039 -0.02051299 -1.05985535 -0.2311598
#> [6,] -1.03874778 -0.3121822 -0.1645355 -0.14097307  1.33026232 -1.7632391
#>              Z19         Z20
#> [1,]  1.13197467  0.05459292
#> [2,] -0.26668710  1.21362397
#> [3,] -0.09568543  0.50513811
#> [4,] -3.22213593 -0.66625942
#> [5,] -0.67129623 -0.99512611
#> [6,]  0.93078885 -0.91234266